requirements.txt
pdfplumber>=0.11.0
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.30.0
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.0
thefuzz>=0.20.0
python-Levenshtein>=0.21.0
pandas>=2.0.0
xlsxwriter>=3.1.0
cryptography>=41.0.0
PyJWT>=2.8.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
boto3
pytz
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
